ChemoSpec 🟠 ➡️ 🔵 hyperSpec

ChemoSpec 🟠 ➡️ 🔵 hyperSpec

Usage

ChemoSpec_to_hyperSpec(Spectra, ...)

Arguments

Spectra

ChemoSpec::Spectra() object.

...

Arguments passed on to hyperSpec::hyperSpec

gc

(logical)
Use garbage collection. If the option gc is set to TRUE, the initialization will have frequent calls to base::gc(), which can help avoid swapping or running out of memory. The default value of gc can be set via hy_set_options().

Value

A hyperSpec::hyperSpec object.

See also

Examples

suppressPackageStartupMessages(library(ChemoSpec))
data("alignMUD")
ChemoSpec_to_hyperSpec(alignMUD) |> str()
#> Warning: replacing previous import ‘lazyeval::is_lang’ by ‘rlang::is_lang’ when loading ‘hyperSpec’
#> Warning: replacing previous import ‘lazyeval::call_modify’ by ‘rlang::call_modify’ when loading ‘hyperSpec’
#> Warning: replacing previous import ‘lazyeval::is_call’ by ‘rlang::is_call’ when loading ‘hyperSpec’
#> Warning: replacing previous import ‘lazyeval::is_formula’ by ‘rlang::is_formula’ when loading ‘hyperSpec’
#> Warning: replacing previous import ‘lazyeval::as_name’ by ‘rlang::as_name’ when loading ‘hyperSpec’
#> Warning: replacing previous import ‘lazyeval::f_label’ by ‘rlang::f_label’ when loading ‘hyperSpec’
#> Warning: replacing previous import ‘lazyeval::is_atomic’ by ‘rlang::is_atomic’ when loading ‘hyperSpec’
#> Warning: replacing previous import ‘lazyeval::is_pairlist’ by ‘rlang::is_pairlist’ when loading ‘hyperSpec’
#> Warning: replacing previous import ‘lazyeval::call_standardise’ by ‘rlang::call_standardise’ when loading ‘hyperSpec’
#> Warning: replacing previous import ‘lazyeval::expr_text’ by ‘rlang::expr_text’ when loading ‘hyperSpec’
#> Warning: replacing previous import ‘lazyeval::f_env<-’ by ‘rlang::f_env<-’ when loading ‘hyperSpec’
#> Warning: replacing previous import ‘lazyeval::f_text’ by ‘rlang::f_text’ when loading ‘hyperSpec’
#> Warning: replacing previous import ‘lazyeval::missing_arg’ by ‘rlang::missing_arg’ when loading ‘hyperSpec’
#> Warning: replacing previous import ‘lazyeval::f_rhs<-’ by ‘rlang::f_rhs<-’ when loading ‘hyperSpec’
#> Warning: replacing previous import ‘lazyeval::f_rhs’ by ‘rlang::f_rhs’ when loading ‘hyperSpec’
#> Warning: replacing previous import ‘lazyeval::f_env’ by ‘rlang::f_env’ when loading ‘hyperSpec’
#> Warning: replacing previous import ‘lazyeval::f_lhs<-’ by ‘rlang::f_lhs<-’ when loading ‘hyperSpec’
#> Warning: replacing previous import ‘lazyeval::expr_label’ by ‘rlang::expr_label’ when loading ‘hyperSpec’
#> Warning: replacing previous import ‘lazyeval::f_lhs’ by ‘rlang::f_lhs’ when loading ‘hyperSpec’
#> Warning: replacing previous import ‘rlang::as_list’ by ‘xml2::as_list’ when loading ‘hyperSpec’
#> Warning: replacing previous import ‘latticeExtra::layer’ by ‘ggplot2::layer’ when loading ‘hyperSpec’
#> Formal class 'hyperSpec' [package "hyperSpec"] with 4 slots
#>   ..@ wavelength: num [1:1500] 5 5 4.99 4.99 4.99 ...
#>   ..@ data      :'data.frame':	20 obs. of  5 variables:
#>   .. ..$ groups : Factor w/ 1 level "S": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ colors : chr [1:20] "black" "black" "black" "black" ...
#>   .. ..$ sym    : num [1:20] 19 19 19 19 19 19 19 19 19 19 ...
#>   .. ..$ alt.sym: chr [1:20] "+" "+" "+" "+" ...
#>   .. ..$ spc    : num [1:20, 1:1500] 0.01225 0.01043 0.00955 0.01113 0.01042 ...
#>   .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. ..$ : NULL
#>   .. .. .. ..$ : chr [1:1500] "5" "4.99666" "4.99333" "4.98999" ...
#>   ..@ label     :List of 2
#>   .. ..$ .wavelength:  expression("ppm")
#>   .. ..$ spc        :  expression("intensity")
#>   ..@ log       :'data.frame':	0 obs. of  0 variables
#> Formal class 'data.frame' [package "methods"] with 4 slots
#>   .. .. ..@ .Data    : list()
#>   .. .. ..@ names    : chr(0) 
#>   .. .. ..@ row.names: int(0) 
#>   .. .. ..@ .S3Class : chr "data.frame"